Accumulo-41 - 4 spaces beat two spaces, 160 line wrap beat no line wrap. These can be opened for discussion later, but for now that is the standard. The eclipse formatter XML file is located in contrib, so no one can use the excuse that they don't have access to it or that it's not documented.
For note- Do not autoformat the master without fixing the little state diagram for transitionOK[][]. Also, no thrift code is reformated since it's autogenerated. No non-.java class files have been modified because the formatter does not have the same type of behavior for those files as far as I could tell.

git-svn-id: https://svn.apache.org/repos/asf/incubator/accumulo/branches/1.3@1186455 13f79535-47bb-0310-9956-ffa450edef68

+++ src/server/src/test/java/org/apache/accumulo/server/tabletserver/log/MultiReaderTest.java
@@ -1,19 +1,19 @@
 /*
-* Licensed to the Apache Software Foundation (ASF) under one or more
-* contributor license agreements.  See the NOTICE file distributed with
-* this work for additional information regarding copyright ownership.
-* The ASF licenses this file to You under the Apache License, Version 2.0
-* (the "License"); you may not use this file except in compliance with
-* the License.  You may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.apache.accumulo.server.tabletserver.log;
 
 import static org.junit.Assert.*;
@@ -33,7 +33,6 @@ import org.apache.log4j.Logger;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-
 
 public class MultiReaderTest {
     
@@ -61,13 +60,12 @@ public class MultiReaderTest {
         }
         writer.close();
     }
-
+    
     @After
     public void tearDown() throws Exception {
-        if (fs != null)
-            fs.delete(new Path("manyMaps"), true);
+        if (fs != null) fs.delete(new Path("manyMaps"), true);
     }
-
+    
     private void scan(MultiReader reader, int start) throws IOException {
         IntWritable key = new IntWritable();
         BytesWritable value = new BytesWritable();
@@ -89,15 +87,13 @@ public class MultiReaderTest {
         }
     }
     
-    
     @Test
     public void testMultiReader() throws IOException {
         MultiReader reader = new MultiReader(fs, conf, "manyMaps");
         IntWritable key = new IntWritable();
         BytesWritable value = new BytesWritable();
         
-        for (int i = 0; i < 1000; i++)
-        {
+        for (int i = 0; i < 1000; i++) {
             if (i == 10) continue;
             assertTrue(reader.next(key, value));
             assertEquals(i, key.get());
@@ -137,5 +133,5 @@ public class MultiReaderTest {
         reader.close();
         
     }
-
+    
 }
